packages/llm_gateway/pyproject.toml
[project]
name = "llm-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Bhaskar Mahajan", email = "bhaskarmahajan98@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.2.2",
    "vllm>=0.18.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "pytest-repeat>=0.9.4"
]

[project.scripts]
llm-gateway = "llm_gateway:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "unit: fast mocked tests for client behavior",
    "smoke: live provider smoke tests that require API credentials",
]

[build-system]
requires = ["uv_build>=0.9.1,<0.10.0"]
build-backend = "uv_build"
